requirements.txt
Django==5.0.3
djangorestframework==3.14.0
python-decouple==3.8
